requirements.txt
pandas
numpy
scikit-learn
faiss-cpu
sentence-transformers
ollama
fastapi
uvicorn
tqdm
matplotlib
seaborn
reportlab